Correctly show scalable and unthemed icons
Previously we were ignoring scalable icons but it seems like the Hicolor
theme puts big icons as scalable even though they are in directories
with the same naming conventions as the ones that are considered non
scalable. Also, the elementary theme sets all directories as scalable.

Because of these cases we can't simply ignore scalable icons. But I'm
still not sure how to show this in the GUI. Maybe the best approach is
to show all icons and on click show the directory information.

Finally the special case for unthemed icons wasn't implemented so we
were creashing on that case. This case is now correctly handled.