Attempt to solve the undefined meaning of scalability
There is a discrepancy between what app developers think are doing when
installing icons into directories and what the icon theme specification
actually says on how they are interpreted.

Before we took the side of the specification and displayed when it
thought an icon was scalable. Now we side with app developers with a
much simpler semantics. We'll see how it works out.