Store icon themes in a linked list not an array
This removes some complexity like a dynamic array implementation for
icon_theme_t pointers. Solves some ownership issues with icon_theme_t,
now they own themselves, are bootrapped and need to be destroyed if one
is being replaced. But this should not be an issue as they will have the
same lifespan as the application.