Don't try to pretend the "All" theme is a icon_theme_t
The only thing we need of a real theme is the pool and the hash table
that stores the icon names list. We now just store these in app_t, to
make it clear "All" is not really a theme.

This also allows to experiment with other data structures to compare
performance, because the "All" theme is making things horrybly slow.