Try using a Gtree to store all icon names
A tree should be faster as we don't need to sort keys every time we
access them to create icon_list. I implemented it and added timing
probes to test this was actually an improvement. Results and analysis
can be found in timing_stats.txt.

Although it does make things faster, it's not the main issue, something
makes creating the GtkListBox for icon_list horribly slow (+2s).