Move "All" theme widget creation into app initialization
Instead of creating the big GtkListBox for all icon themes every time
the All theme is chosen, do this once at startup and then just
replace it and show it when required.

I also tested not using GtkListBox and just adding a GtkBox with labels
to see if the the slowness was caused by GtkListBox or by the ~7000
widgets we were rendering. Even though things improved, we still took 1s
to switch to the All theme, which is still unacceptable.

Further testing showed that the slow parts were the process of building
the widget, and then the call to gtk_widget_show_all(), using
gtk_widget_show() for all rows and then calling gtk_widget_show() on the
widged improved things, but again, not by much. Calling
gtk_widget_show() once at app startup, and then just using
gtk_container_add() when replacing it made the 1s slowness of the
showing process move into the adding process.

Also the call to gtk_container_remove() when changing the All theme for
another one, took about 1s.

A theory I found on the gtk mailing list about this is that a LOT of
signals are thrown whe adding this much widgets, the best solution seems
to simply not use this much widgets.

With this change now the application takes 2s to open (creating the
widget), about 1s then de All theme is selected (the call to
gtk_widget_show_all) and about another second to change from the All
theme (call to getk_container_remove). I expect this to be fixed in the
next commits.