Simplify replace_wrapped_widget_defered()
Most of the work must be done before the destruction of the widget,
otherwise multiple fast calls to replace_wrapped_widget_defered() won't
be handled properly. Although showing the new widget has to be done here
to avoid glitching and extending the wrapper's size unnecessarily.

Also fixed a unitialized use of idx when using -O3.